dify/tests/requirements-test.txt
# 测试依赖

# 测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Mock 工具
pytest-mock>=3.12.0

# 代码覆盖率
coverage>=7.3.0

# 环境变量
python-dotenv>=1.0.0

# HTTP客户端 (如果主项目没有)
httpx>=0.24.0
